backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
motor==3.6.0
pydantic==2.10.4
python-dotenv==1.0.1
starlette==0.41.3
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
bcrypt==4.2.1
requests==2.31.0